add rellume example
rellume seems to take byte code, in the example it is on
the level of a function, and compile to llvm ir. This might
be too small/scoped for what we want to do, but could still
offer a means to test something out, so worth including.
The container added here (pushed with tag "rellume" builds
the software and examples

Signed-off-by: vsoch <vsoch@users.noreply.github.com>